+++ jxls-core/src/test/java/net/sf/jxls/CellsChecker.java
@@ -1,17 +1,16 @@
 package net.sf.jxls;
 
+import junit.framework.Assert;
+import org.apache.poi.hssf.usermodel.HSSFCell;
+import org.apache.poi.hssf.usermodel.HSSFCellStyle;
+import org.apache.poi.hssf.usermodel.HSSFRow;
+import org.apache.poi.hssf.usermodel.HSSFSheet;
+
 import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
-
-import junit.framework.Assert;
-
-import org.apache.poi.hssf.usermodel.HSSFCell;
-import org.apache.poi.hssf.usermodel.HSSFCellStyle;
-import org.apache.poi.hssf.usermodel.HSSFRow;
-import org.apache.poi.hssf.usermodel.HSSFSheet;
 
 /**
  * @author Leonid Vysochyn
@@ -171,8 +170,8 @@ public class CellsChecker extends Assert {
         }
     }
 
-    static short getWidth(HSSFSheet sheet, short col) {
-        short width = sheet.getColumnWidth(col);
+    static int getWidth(HSSFSheet sheet, int col) {
+        int width = sheet.getColumnWidth(col);
         if (width == sheet.getDefaultColumnWidth()) {
             width = (short) (width * 256);
         }
