+++ jxls-reader/src/main/java/net/sf/jxls/reader/SimpleBlockReaderImpl.java
@@ -1,16 +1,16 @@
 package net.sf.jxls.reader;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.poi.hssf.usermodel.HSSFCell;
 import org.apache.poi.hssf.usermodel.HSSFRow;
 import org.apache.poi.hssf.usermodel.HSSFSheet;
 import org.apache.poi.hssf.util.CellReference;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author Leonid Vysochyn
@@ -133,7 +133,7 @@ public class SimpleBlockReaderImpl extends BaseBlockReader implements SimpleBloc
         return beanCellMappings;
     }
 
-    private HSSFCell getCell(HSSFSheet sheet, int rowNum, short cellNum) {
+    private HSSFCell getCell(HSSFSheet sheet, int rowNum, int cellNum) {
         HSSFRow row = sheet.getRow(rowNum);
         if (row == null) {
             return null;
